OviPro/MundoCampero/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
fonttools==4.55.3
idna==3.10
numpy==2.0.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
wkhtmltopdf==0.2
XlsxWriter==3.2.0
zopfli==0.2.3.post1
